requirements.txt
arrow==1.3.0
asgiref==3.8.1
autopep8==2.3.2
binaryornot==0.4.4
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.1
cssselect2==0.7.0
debugpy==1.8.7
defusedxml==0.7.1
Django==5.1.5
django-allauth==65.0.2
django-environ==0.12.0
django-paypal==2.1
django-tailwind==3.8.0
fonttools==4.55.3
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.2
paypalrestsdk==1.13.3
pillow==11.0.0
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyotp==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20241206
tzdata==2024.1
urllib3==2.2.3
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
